la barra de navegación es compartida por todas las pantallas y ya se puede cambiar el username. Eliminé la opción de cambiar la contraseña
