+++ lib/users.dart
@@ -13,4 +13,18 @@ class User {
   bool notificaciones = true;
 
   User({required this.username, required this.email, required this.password, required this.pets, required this.zones, required List<Reporte> reportes});
+
+  Map<String, dynamic> toMap() {
+    return {
+      'username': username,
+      'email': email,
+      'password': password,
+      'pets': pets.map((pet) => pet.toMap()).toList(),
+      'reportes': reportes.map((reporte) => reporte.toMap()).toList(),
+      'zones': zones,
+      'icon': icon,
+      'notificaciones': notificaciones,
+    };
+  }
+
 }
